Refactor frontend celebration content
